requirements.txt
json-spec (0.9.8)
pip (1.5.4)
Pygments (1.6)
setuptools (2.2)
suds-jurko (0.6)
Werkzeug (0.9.6)
